requirements.txt
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
azure-storage-blob==12.17.0